Kernel: Split the Plan9FileSystem.{cpp,h} file into smaller components
