+++ Kernel/FileSystem/Plan9FS/Definitions.h
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2020, Sergey Bugaev <bugaevc@serenityos.org>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <AK/Types.h>
+
+namespace Kernel {
+
+struct Plan9FSQIdentifier {
+    u8 type;
+    u32 version;
+    u64 path;
+};
+
+}
